Camera resets before entering the game
Saved games also store the camera position.

+++ src/main/java/com/bitbucket/computerology/gui/elements/GameCanvas.java
@@ -13,7 +13,7 @@ import org.newdawn.slick.Graphics;
 public class GameCanvas extends GUIElement {
     
     boolean dragging = false;
-    int last_x = 0, last_y = 0;
+    double last_x = 0, last_y = 0;
     
     public GameCanvas() {
